Return suggestions when the $type is not set on SpatieTagsInput
If there is no type set on the SpatieTagsInput component and tags records have types, the suggestions do not work. This is because the suggestions query adds a condition that a type should be NULL if it is not set. By removing this condition, the suggestions always return all types if no type is set. I believe this is the correct behaviour.